added motion controls
. I added controls to all sprites (see screenshot on discord)

2. I added parameters for different animations to happen. They are
isShaking which is boolean -- I added an "Accelerometer" script which we may use to get this information. perhaps it is redundant and needs to be in each sprite's movement script.
This will cause the sprites to go into "oshit" mode when it is true.
This will cause the sprites to go back to normal when it is false.

isNormal another boolean -- this will cause the sprites to go from "back_to_normal" animation to default which would be swimming.

touch this is a trigger -- this will need to be a script that says if a bubble is touched which will then trigger the pop animation

Another script will likely be needed to send up bubbles at randomized intervals, but I suspect isShaking will need to be used here to determine frequency.

I haven't touched sounds yet.